+++ R/global_vars.R
@@ -208,7 +208,12 @@ utils::globalVariables(
     "95% ci.lb",
     "95% ci.ub",
     "z value",
-    "prior"
+    "prior",
+    "Est",
+    "ci.low",
+    "ci.up",
+    "effect_size",
+    "df_internal"
   ),
   package = "statsExpressions",
   add = FALSE