+++ R/reexports.R
@@ -44,6 +44,10 @@ rlang::`%||%`
 rlang::`%|%`
 
 #' @export
+#' @importFrom rlang ":="
+rlang::`:=`
+
+#' @export
 #' @importFrom ipmisc tibble
 ipmisc::tibble