+++ tests/testthat/test-expr_meta_parametric.R
@@ -98,36 +98,6 @@ testthat::test_that(
     testthat::expect_identical(using_function1, results1)
     testthat::expect_identical(using_function2, results2)
 
-    # tidy dataframe
-    set.seed(123)
-    tidy_df <-
-      expr_meta_parametric(
-        data = df_eg,
-        messages = FALSE,
-        output = "tidy"
-      )
-
-    glance_df <-
-      expr_meta_parametric(
-        data = df_eg,
-        messages = FALSE,
-        output = "glance"
-      )
-
-    # checking if the tidy output is expected
-    testthat::expect_equal(tidy_df$estimate, 0.4376927, tolerance = 0.0001)
-    testthat::expect_equal(tidy_df$conf.low, 0.04231262, tolerance = 0.0001)
-    testthat::expect_equal(tidy_df$conf.high, 0.8330728, tolerance = 0.0001)
-
-    # checking if the glance output is expected
-    testthat::expect_equal(dim(glance_df), c(1L, 12L))
-    testthat::expect_equal(glance_df$tau2, 0.193525, tolerance = 0.0001)
-    testthat::expect_equal(glance_df$se.tau2, 0.1437787, tolerance = 0.0001)
-    testthat::expect_equal(glance_df$QE, 6082.595, tolerance = 0.01)
-    testthat::expect_equal(glance_df$k, 5L)
-    testthat::expect_equal(glance_df$p, 1L)
-    testthat::expect_equal(glance_df$m, 1L)
-
     # error
     testthat::expect_error(expr_meta_parametric(mtcars))
   }